refactor(mobile): move memberships + seen_state to the modules that own them
Mobile kept memberships and seen_state in collections/admin.ts while web kept
them in organization and communication — the same table reached through a
different module depending on which app you were reading. Align mobile to web.

The placement is not a taste call: membershipTable is declared in
organization-tables.ts and seenStateTable in communication-tables.ts, and
domain-types puts Membership in organization.ts. Web already matched the schema;
mobile was the outlier. admin.ts is now users only, mirroring web's users+teams.

Lifecycle is preserved exactly. initializeMembershipsCollection moves to
organization and initializeSeenStateCollection to communication, each keeping its
own initializer rather than being folded into the scope-driven ones — seen state
is scoped `user_id = me` server-side and must NOT rebuild on a channel-set
resync, which is why it stays out of initializeCommunicationCollections. The
teardowns move into the matching reset functions; all three are still called from
resetAllCollections, which only runs at sign-out, so both collections keep the
same single teardown point they had before.

Prerequisite for sharing the collection descriptors themselves (#2): the two
apps have to agree on which module a collection lives in before the specs can be
lifted into one place.

Behavior-preserving. Typecheck holds at the 85-error baseline, 18 tests pass, and
executor.ts's duplicate-import warning (introduced by the move) is fixed.
